+++ src/components/POLine/ChangeInstanceConnection/hooks/index.js
@@ -1,0 +1,2 @@
+export { useChangeInstanceConnection } from './useChangeInstanceConnection';
+export { useChangeInstanceModalConfigs } from './useChangeInstanceModalConfigs';